Move indexing missing odc cli to seperate modules
