+++ setup.cfg
@@ -52,11 +52,11 @@ console_scripts =
     s1-gap-report = deafrica.monitoring.s1_gap_report:cli
     check-dead-queue = deafrica.monitoring.check_dead_queues:cli
     s2-gap-report = deafrica.monitoring.s2_gap_report:cli
-    s2-gap-filler = deafrica.monitoring.s2_gap_filler:s2_gap_filler_cli
-    s2-missing-odc-filler = deafrica.monitoring.s2_gap_filler:s2_missing_odc_cli
+    s2-gap-filler = deafrica.monitoring.s2_gap_filler:cli
+    s2-index-missing-odc = deafrica.monitoring.s2_index_missing_odc:cli
     s2-c1-gap-report = deafrica.monitoring.s2_c1_gap_report:cli
-    s2-c1-gap-filler = deafrica.monitoring.s2_c1_gap_filler:s2_c1_gap_filler_cli
-    s2-c1-missing-odc-filler =  deafrica.monitoring.s2_c1_gap_filler:s2_c1_missing_odc_cli
+    s2-c1-gap-filler = deafrica.monitoring.s2_c1_gap_filler:cli
+    s2-c1-index-missing-odc =  deafrica.monitoring.s2_c1_index_missing_odc:cli
     landsat-gap-report = deafrica.monitoring.landsat_gap_report:cli
     landsat-gap-filler = deafrica.monitoring.landsat_gap_filler:cli
     latency-check = deafrica.monitoring.latency_check:cli
